+++ src/utils/media.util.ts
@@ -1,0 +1,62 @@
+export const SoundSourceMapping = {
+  Typing: './assets/sounds/typing.mp3',
+  Correct: './assets/sounds/correct.wav',
+  Incorrect: './assets/sounds/incorrect.wav',
+} as const;
+
+export type SoundsType = keyof typeof SoundSourceMapping;
+
+export const BANNED_KEYS = [
+  'Enter',
+  'Backspace',
+  'Delete',
+  'Tab',
+  'CapsLock',
+  'Shift',
+  'Control',
+  'Alt',
+  'Meta',
+  'Escape',
+  'Fn',
+  'FnLock',
+  'Hyper',
+  'Super',
+  'OS',
+  // Up, down, left and right keys
+  'ArrowUp',
+  'ArrowDown',
+  // volume keys
+  'AudioVolumeUp',
+  'AudioVolumeDown',
+  'AudioVolumeMute',
+  // special keys
+  'End',
+  'PageDown',
+  'PageUp',
+  'Clear',
+  'Home',
+  '1',
+  '2',
+  '3',
+  '4',
+  '5',
+  '6',
+  '7',
+  '8',
+  '9',
+  '0',
+];
+
+export const isChineseSymbol = (val: string): boolean =>
+  /[\u3002|\uff1f|\uff01|\uff0c|\u3001|\uff1b|\uff1a|\u201c|\u201d|\u2018|\u2019|\uff08|\uff09|\u300a|\u300b|\u3008|\u3009|\u3010|\u3011|\u300e|\u300f|\u300c|\u300d|\ufe43|\ufe44|\u3014|\u3015|\u2026|\u2014|\uff5e|\ufe4f|\uffe5]/.test(
+    val,
+  );
+
+export const playSound = (params: { soundsType?: SoundsType; src?: string }) => {
+  const { soundsType, src } = params;
+  const audio = new Audio();
+  const audioUrl = src || SoundSourceMapping[soundsType];
+  audio.src = audioUrl;
+  audio.load();
+  audio.play();
+};
